채팅방_화면_개선사항_수정 : feat : 시간관련 util 함수 common_utils로 이동, trim연산 중복 방지, subscribeToUserMessages 삭제 , common appBar 기본뒤로가기 값 변경 삭제 https://github.com/TEAM-ROMROM/RomRom-FE/issues/363
